rename state_changes file + create LakeConfig + export types for near-lake-framework
