+++ packages/near-lake-primitives/src/types/index.ts
@@ -1,6 +1,10 @@
-export { Block } from './block';
+export { LakeContext } from './lakeContext';
+
+export { Block, streamerMessageToBlock } from './block';
 export { Event } from './events';
 export { Receipt } from './receipts';
-export { StateChange } from './state_changes';
+export { StateChange } from './stateChanges';
 export { Transaction } from './transactions';
 export * from './core'
+
+
