+++ packages/near-lake-primitives/src/types/lakeContext.ts
@@ -1,0 +1,3 @@
+export class LakeContext {
+    // TODO
+}
